Eager Loading
Learned about the difference between eager and lazing loading:
Eager loading is loading all the associated entities at once(example, when retrieving instructor, all the courses associated with the retrieved instructor will be retrieved as well). Lazy loading is loading the main entities only and loading the associated entities only on-demand.
It is preferable to use lazy loading as it enhances the performance in case of large number of data.

- Tested eager loading by adding fetch type to the list of courses
-Tested in DEBUG mode to see what entities are retrieved